Create separate package for imagesources
