+++ volumina/pixelpipeline/imagesources/__init__.py
@@ -1,7 +1,3 @@
-from __future__ import print_function
-from __future__ import absolute_import
-from __future__ import division
-
 ###############################################################################
 #   volumina: volume slicing and editing library
 #
@@ -23,7 +19,6 @@ from __future__ import division
 # This information is also available on the ilastik web site at:
 # 		   http://ilastik.org/license/
 ###############################################################################
-# Python
 from past.utils import old_div
 import logging
 import time
@@ -39,6 +34,7 @@ from volumina.slicingtools import is_bounded, slicing2rect, rect2slicing, slicin
 from volumina.config import CONFIG
 from volumina.utility import execute_in_main_thread
 import numpy as np
+from ._base import ImageSource
 
 _has_vigra = True
 try:
@@ -71,66 +67,6 @@ def log_request(func):
 # *******************************************************************************
 # I m a g e S o u r c e                                                        *
 # *******************************************************************************
-
-
-class ImageSource(QObject, ImageSourceABC):
-    """Partial implemented base class for image sources
-
-    Signals:
-    isDirty -- a rectangular region has changed; transmits
-               an empty QRect if the whole image is dirty
-
-    """
-
-    isDirty = pyqtSignal(QRect)
-
-    def __init__(self, name, guarantees_opaqueness=False, parent=None, direct=False):
-        """ direct: whether this request will be computed synchronously in the GUI thread (direct=True)
-                    or whether the request will be put on a worker queue to be computed in a worker thread
-                    (direct=False).
-                    Only use direct=True if the layer's data will be immediately available"""
-        super(ImageSource, self).__init__(parent=parent)
-        self._opaque = guarantees_opaqueness
-        self.direct = direct
-        self.name = name
-
-    def image_type(self):
-        """
-        Image sources must declare what type of "image" they will produce.
-        The two allowed types are QImage and QGraphicsItems (and subclasses).
-        """
-        return QImage
-
-    def request(self, rect, along_through=None):
-        raise NotImplementedError
-
-    def setDirty(self, slicing):
-        """Mark a region of the image as dirty.
-
-        slicing -- if one ore more slices in the slicing
-                   are unbounded, the whole image is marked dirty;
-                   since an image has two dimensions, only the first
-                   two slices in the slicing are used
-
-        """
-        if not is_pure_slicing(slicing):
-            raise Exception("dirty region: slicing is not pure")
-        if not is_bounded(slicing):
-            self.isDirty.emit(QRect())  # empty rect == everything is dirty
-        else:
-            self.isDirty.emit(slicing2rect(slicing))
-
-    def isOpaque(self):
-        """Image is opaque everywhere (i.e. no pixel has an alpha value != 255).
-
-        If the ImageSource can give an opaqueness guarantee,
-        performance can be improved since layers occluded by this
-        source don't have to be rendered in some cases.
-
-        Warning: Can cause rendering bugs: In doubt return False.
-
-        """
-        return self._opaque
 
 
 # *******************************************************************************
